+++ EduBot/src/test/java/xbot/edubot/TogglePrecisionDriveCommandTest.java
@@ -11,45 +11,45 @@ import xbot.common.controls.sensors.mock_adapters.MockJoystick;
 
 public class TogglePrecisionDriveCommandTest extends BaseDriveTest {
 
-	@Test
-	public void test() {
-		DriveSubsystem drive = this.injector.getInstance(DriveSubsystem.class);
-		OperatorInterface oi = this.injector.getInstance(OperatorInterface.class);
-		
-		BaseCommand driveCommand = new TankDriveWithJoysticksCommand(drive, oi);
-		BaseCommand togglePrecisionCommand = new TogglePrecisionDriveCommand(drive);
-		
-		MockJoystick left = (MockJoystick)oi.leftJoystick;
-		MockJoystick right = (MockJoystick)oi.rightJoystick;
-		
-		driveCommand.initialize();
-		
-		togglePrecisionCommand.initialize();
-		
-		if(!togglePrecisionCommand.isFinished()){
-			togglePrecisionCommand.execute();
-		}
-		left.setY(1.0);
-		right.setY(1.0);
-		driveCommand.execute();
-		this.assertDrive(0.5, 0.5);
-		
-		if(!togglePrecisionCommand.isFinished()){
-			togglePrecisionCommand.execute();
-		}
-		left.setY(-1.0);
-		right.setY(-1.0);
-		driveCommand.execute();
-		this.assertDrive(-0.5, -0.5);
-		
-		togglePrecisionCommand.initialize();
-		if(!togglePrecisionCommand.isFinished()){
-			togglePrecisionCommand.execute();
-		}
-		left.setY(1.0);
-		right.setY(1.0);
-		driveCommand.execute();
-		this.assertDrive(1, 1);
-		
-	}
+    @Test
+    public void test() {
+        DriveSubsystem drive = this.injector.getInstance(DriveSubsystem.class);
+        OperatorInterface oi = this.injector.getInstance(OperatorInterface.class);
+        
+        BaseCommand driveCommand = new TankDriveWithJoysticksCommand(drive, oi);
+        BaseCommand togglePrecisionCommand = new TogglePrecisionDriveCommand(drive);
+        
+        MockJoystick left = (MockJoystick)oi.leftJoystick;
+        MockJoystick right = (MockJoystick)oi.rightJoystick;
+        
+        driveCommand.initialize();
+        
+        togglePrecisionCommand.initialize();
+        
+        if(!togglePrecisionCommand.isFinished()){
+            togglePrecisionCommand.execute();
+        }
+        left.setY(1.0);
+        right.setY(1.0);
+        driveCommand.execute();
+        this.assertDrive(0.5, 0.5);
+        
+        if(!togglePrecisionCommand.isFinished()){
+            togglePrecisionCommand.execute();
+        }
+        left.setY(-1.0);
+        right.setY(-1.0);
+        driveCommand.execute();
+        this.assertDrive(-0.5, -0.5);
+        
+        togglePrecisionCommand.initialize();
+        if(!togglePrecisionCommand.isFinished()){
+            togglePrecisionCommand.execute();
+        }
+        left.setY(1.0);
+        right.setY(1.0);
+        driveCommand.execute();
+        this.assertDrive(1, 1);
+        
+    }
 }
