+++ EduBot/src/test/java/xbot/edubot/linear/LinearVisualizationPanel.java
@@ -11,7 +11,7 @@ import javax.swing.JPanel;
 import xbot.common.math.XYPair;
 
 public class LinearVisualizationPanel extends JPanel {
-	private int centerX;
+    private int centerX;
     private int centerY;
     private String loops = "0";
     private double robotCurrentDistance = 0;
@@ -29,12 +29,12 @@ public class LinearVisualizationPanel extends JPanel {
     }
     
     public void setRobotDistance(double distance) {
-    	robotCurrentDistance = distance;
+        robotCurrentDistance = distance;
     }
     
     public void updateViz(DriveToPositionCommandTest.LinearEnvironmentState envState) {
-    	robotCurrentDistance = envState.distance;
-    	loops = Integer.toString(envState.loops);
+        robotCurrentDistance = envState.distance;
+        loops = Integer.toString(envState.loops);
     }
     
     @Override
